// src/app/shared/generated/model/themeDTO.ts
/**
 * tkit-portal-server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 4.4.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PortalIdentifierDTO } from './portalIdentifierDTO'

export interface ThemeDTO {
  version?: number
  creationDate?: string
  creationUser?: string
  modificationDate?: string
  modificationUser?: string
  id?: string
  name: string
  cssFile?: string
  description?: string
  assetsUrl?: string
  logoUrl?: string
  faviconUrl?: string
  previewImageUrl?: string
  assetsUpdateDate?: string
  properties?: object
  portals?: Array<PortalIdentifierDTO>
}